Python 3: Fix unused-argument error
Since the member function `branchinspect(self, branch)` seems to fulfill
an interface, but the `branch` argument is unused.

Instead of replacing the function signature with
`branchinspect(self, _)`, it might be better to simply delete the local
variable inside the function.
This way someone who relies on the argument name of `branchinspect` will
not have their code break when running this code.